Renamed module to channels.security.websockets
